chelalBackend/requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
gunicorn
celery
redis
channels
channels-redis
pytest
pytest-django
requests
twilio
drf-nested-routers
django-cors-headers
django-filter
django-extensions
django-storages
boto3
django-redis
django-axes
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow
Faker

